requirements.txt
altair==5.5.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.2
coverage==7.10.3
debugpy==1.8.12
decorator==5.1.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
findspark==2.0.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==2.0.1
nest-asyncio==1.6.0
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.6
plotly==6.2.0
pluggy==1.6.0
prompt_toolkit==3.0.50
protobuf==6.31.1
psutil==6.1.1
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==19.0.0
pydeck==0.9.1
Pygments==2.19.1
pyspark==3.5.4
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.48.0
tenacity==9.1.2
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.1
urllib3==2.5.0
wcwidth==0.2.13
